refactor: app.pyのChatGPTの処理をmodules/chatgpt.pyに移動
